fix(ci): compatibility with Quarkus 3.16.0.CR1
QuarkusUnitTest 3.16+ contains a change of behavior: it only allows 1 properties file per test (the last one wins).
This was detected by the nightly build based on Quarkus master (999-SNAPSHOT), but can now also be reproduced with the recently released 3.16.0.CR1.
To avoid branching for this, it's simpler to refactor the tests to work both with Quarkus 3.15 and 3.16.