Include HttpEventCollectorResendMiddleware from PR to properly implement retrying
As of today the 'bundled' HttpEventCollectorResendMiddleware only retries when '200 - OK' response is returned from the Splunk server, which is odd to say the least. Simon Hege has created a PR on the (https://github.com/splunk/splunk-library-javalogging/pull/287)[main java splunk logging project], but this project seems to be inactive to the point that no contributor is merging this. I suggest to include this in the Quarkus logger, until https://github.com/quarkiverse/quarkus-logging-splunk/issues/281 is implemented.