fix: add option to disable printing log events to stdout
Introduces a new configuration option to disable the printing of
log events to stdout when a communication error with splunk occurs.
This is useful in scenarios where applications manages sensitive data
such as PCI, PII or security token.
Also fix issue build by adding a asciidoctorj-diagram-plantuml dependency
due to asciidoctorj-diagram to 3.0.1